# references/swagger_external.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Open API ONCall v3 Externals",
    "description": "Open API for ONCall and Mechanical cases.",
    "version": "2025-08-12T17:37:48.000Z"
  },
  "servers": [
    {
      "url": "https://qualif.api.michelin.com"
    },
    {
      "url": "https://api.michelin.com"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/amn/moc/v3/cases/{id}": {
      "get": {
        "tags": [
          "Cases"
        ],
        "description": "Get unified case (v3)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified case number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "created_at": "2025-07-27T20:11:52.211Z",
                    "updated_at": "2025-08-03T19:24:02.776Z",
                    "customerST": "1471852",
                    "customerBT": "1424651",
                    "id": "Z8D13493LK",
                    "ersResponse": {
                      "data": {
                        "assetLocation": {
                          "address": "202 GSP Logistic Parkway",
                          "city": "Greer",
                          "country": "USA",
                          "note": "comment",
                          "state": "SC",
                          "latitude": 34.9289684,
                          "longitude": -82.2007266,
                          "zipCode": "29651",
                          "highway": null,
                          "milemarker": null,
                          "timezone": null
                        },
                        "etaDuration": {
                          "min": null,
                          "max": null
                        },
                        "fixpix": {
                          "sent": "not_attempted",
                          "sentTimestamp": ""
                        },
                        "caseNumber": "E484128",
                        "caseStatus": "new",
                        "previousStatus": null,
                        "statusTimeStamp": "2025-07-27T20:11:59.712Z",
                        "userName": "maestro@michelin.com",
                        "typeOfCase": "api-call",
                        "inboundProgram": "Michelin OnCall - Electronic Dispatch",
                        "inboundProgramNumberId": 229,
                        "inboundCallerStore": null,
                        "inboundStoreBillTo": "NA",
                        "inboundStoreShipTo": null,
                        "paymentMethod": "michelin_line_of_credit",
                        "customerName": "HUB GROUP TRUCKING, INC. (ERS)",
                        "accountNumber": "1471852",
                        "providerName": null,
                        "poNumber": "888889",
                        "drNumber": "DR123204909",
                        "refNumber": "ref12345",
                        "ticketNumber": "DR123204909",
                        "invoiceNumber": "345",
                        "billingAmount": "0.00",
                        "workOrderNumber": null,
                        "billingComment": "test",
                        "releaseNumber": null,
                        "contractNumber": null,
                        "servicingDealerBilltoName": null,
                        "servicingDealerStore": null,
                        "servicingDealerTimezone": null,
                        "specialInstructions": "test patch",
                        "delayedService": true,
                        "delayedServiceNote": null,
                        "delayedServiceCompletionDate": "2025-07-26T00:00:00.000Z",
                        "delayedServiceDispatchDate": null,
                        "updatedAt": "2025-08-03T19:23:54.518Z",
                        "insertedAt": "2025-07-27T20:11:59.712Z",
                        "billable": false,
                        "customer": {
                          "shipTo": "1471852",
                          "billTo": "1424651",
                          "name": "HUB GROUP TRUCKING, INC. (ERS)",
                          "city": "OAK BROOK",
                          "state": "IL",
                          "zip": "60523",
                          "street": "2000 CLEARWATER DR"
                        },
                        "contacts": [
                          {
                            "contactName": "John",
                            "contactPhone": "+19314090552",
                            "contactType": "dispatch",
                            "contactEmail": null,
                            "callbackEta": false,
                            "callbackRolltime": false
                          },
                          {
                            "contactName": "John",
                            "contactPhone": "+19314090552",
                            "contactType": "driver_cell",
                            "contactEmail": null,
                            "callbackEta": false,
                            "callbackRolltime": false
                          }
                        ],
                        "serviceData": [],
                        "requestedService": [
                          {
                            "type": "request",
                            "unitNumber": "MT2626",
                            "action": "replace",
                            "condition": "blank_case",
                            "position": "rf",
                            "axle": "other",
                            "tireSize": "10.00/90R20",
                            "tireBrand": "Michelin",
                            "treadDesign": "PILOT XFA",
                            "productType": "New",
                            "rimMaterial": null,
                            "wheelOffset": null,
                            "rimType": null,
                            "loadRange": "see_notes",
                            "confirmed": null
                          }
                        ],
                        "agreedService": [
                          {
                            "type": "agreement",
                            "unitNumber": "MT2626",
                            "action": "replace",
                            "condition": "blank_case",
                            "position": "rf",
                            "axle": "other",
                            "tireSize": "10.00/90R20",
                            "tireBrand": "Michelin",
                            "treadDesign": "PILOT XFA",
                            "productType": "New",
                            "rimMaterial": null,
                            "wheelOffset": null,
                            "rimType": null,
                            "loadRange": "see_notes",
                            "confirmed": null
                          }
                        ],
                        "suppliedService": [
                          {
                            "type": "supplied",
                            "unitNumber": "MT2626",
                            "action": "replace",
                            "condition": "blank_case",
                            "position": "rf",
                            "axle": "other",
                            "tireSize": "10.00/90R20",
                            "tireBrand": "Michelin",
                            "treadDesign": "PILOT XFA",
                            "productType": "New",
                            "rimMaterial": null,
                            "wheelOffset": null,
                            "rimType": null,
                            "loadRange": "see_notes",
                            "confirmed": true
                          }
                        ],
                        "caseNotes": [],
                        "linkedCalls": [],
                        "notifications": [],
                        "assets": [
                          {
                            "identifier": "d7ec109b-ac3d-49b6-bb6e-32bcee4364aa",
                            "unitNumber": "MT2626",
                            "unitType": "tractor",
                            "vin": null
                          }
                        ],
                        "customerProfile": {
                          "fleetContactsAvailable": true,
                          "driverAuthorized": null,
                          "poRequired": false,
                          "tirePreferences": true,
                          "wheelPreferences": false,
                          "emailPreferences": true
                        },
                        "caseStatusChanges": [],
                        "delays": []
                      },
                      "codeResponse": 200
                    },
                    "mechanicalResponse": {
                      "data": {
                        "eventNumber": 8837388,
                        "custId": 119758,
                        "customerCode": "MOCKN46",
                        "customerPONumber": "888889",
                        "createDateUtc": "2025-07-27T03:54:22.937Z",
                        "updatedDateUtc": "2025-08-03T20:27:19.75Z",
                        "vendorAssignedDateUtc": "2025-07-27T04:20:00Z",
                        "repairStartedDateUtc": "2025-07-27T06:30:00Z",
                        "repairCompletedDateUtc": "2025-07-27T10:07:00Z",
                        "serviceProviderWorkCompletedDateUtc": null,
                        "status": "rolling"
                      },
                      "codeResponse": 200
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Cases"
        ],
        "description": "Allows to update the case information (billing information, assets, products, location, etc).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Case Number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "billingInfo": {
                    "comment": "Send any comment here",
                    "poNumber": "Po Number here",
                    "drNumber": "DR Number here",
                    "referenceNumber": "reference number here",
                    "invoiceNumber": "invoice number here",
                    "amount": "0.00"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Case update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "Z8D13493LK",
                    "customerST": "1471852",
                    "customerBT": "1424651",
                    "created_at": "2025-07-27T20:11:52.211Z",
                    "updated_at": "2025-08-03T19:24:02.776Z",
                    "ersResponse": {
                      "codeResponse": 200,
                      "data": {
                        "id": "E484128",
                        "inboundProgramNumberId": "229",
                        "status": "NEW",
                        "dispatchedAt": null,
                        "billTo": null,
                        "shipTo": null,
                        "eta": null,
                        "provider": {
                          "name": null,
                          "address": null
                        },
                        "source": "OnCall",
                        "parent_company": null,
                        "home_office": null
                      }
                    },
                    "mechanicalResponse": {
                      "codeResponse": 200,
                      "data": {
                        "eventNumber": 8837388,
                        "custId": 119758,
                        "customerCode": "MOCKN46",
                        "customerPONumber": "888889",
                        "createDateUtc": "2025-07-27T03:54:22.937Z",
                        "updatedDateUtc": "2025-08-03T20:27:19.75Z",
                        "vendorAssignedDateUtc": "2025-07-27T04:20:00Z",
                        "repairStartedDateUtc": "2025-07-27T06:30:00Z",
                        "repairCompletedDateUtc": "2025-07-27T10:07:00Z",
                        "serviceProviderWorkCompletedDateUtc": null,
                        "status": "new"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/idp/v2/b2b/oauth2/token",
            "scopes": {}
          }
        }
      }
    }
  }
}